fix(skills): target canonical workspace from worktree sessions and survive missing drafts
Sessions running in ephemeral worktrees recorded proposal targets under the
worktree, so applied skill updates were silently lost to throwaway copies
(and the workshop later classified its own skills as unowned). The session
producer now records the durable canonical workspace when it swaps cwd into
a managed worktree; capture and the workshop tool prefer that recorded fact,
and absence preserves existing behavior.

A pending create whose durable PROPOSAL.md draft is missing no longer fails
the whole list action: it renders with degradedState=draft-missing, inspect
and apply fail with an actionable message, and reject works draft-free.

Fixes #125653
