+++ src/agents/tools/skill-workshop-tool-factory.ts
@@ -1,6 +1,7 @@
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import type { SkillProposalOrigin, SkillWorkshopRunOptions } from "../../skills/workshop/types.js";
+import { getCanonicalSkillWorkspace } from "../skill-workshop-workspace-context.js";
 import { createSkillWorkshopTool } from "./skill-workshop-tool.js";
 
 export function createConfiguredSkillWorkshopTool(params: {
@@ -18,7 +19,7 @@ export function createConfiguredSkillWorkshopTool(params: {
     params.messageId === undefined ? undefined : String(params.messageId),
   );
   return createSkillWorkshopTool({
-    workspaceDir: params.workspaceDir,
+    workspaceDir: getCanonicalSkillWorkspace() ?? params.workspaceDir,
     config: params.config,
     env: params.run?.env,
     agentId: params.agentId,
