+++ src/skills/workshop/experience-review-auto-apply.test.ts
@@ -1,5 +1,7 @@
 import fs from "node:fs/promises";
+import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { runWithCanonicalSkillWorkspace } from "../../agents/skill-workshop-workspace-context.js";
 import { createSkillWorkshopTool } from "../../agents/tools/skill-workshop-tool.js";
 import {
   isGatewaySubordinateWorkAdmissionClosed,
@@ -10,6 +12,7 @@ import {
   type OpenClawTestState,
 } from "../../test-utils/openclaw-test-state.js";
 import { createTrackedTempDirs } from "../../test-utils/tracked-temp-dirs.js";
+import { writeSkill } from "../test-support/e2e-test-helpers.js";
 import { runSkillExperienceReview, type ExperienceReviewCandidate } from "./experience-review.js";
 import { inspectSkillProposal, listSkillProposals, proposeCreateSkill } from "./service.js";
 
@@ -91,22 +94,18 @@ describe("experience review auto apply", () => {
     );
   });
 
-  it("auto-applies reviewer full-body updates after an authoritative read", async () => {
-    const workspaceDir = await tempDirs.make("openclaw-experience-auto-apply-update-");
-    const seedTool = createSkillWorkshopTool({
-      workspaceDir,
-      config: { skills: { workshop: { approvalPolicy: "auto" } } },
-    });
-    const seeded = await seedTool.execute("seed-create", {
-      action: "create",
+  it("auto-applies updates to the durable workspace from a session worktree", async () => {
+    const canonicalWorkspaceDir = await tempDirs.make("openclaw-experience-canonical-");
+    const worktreeWorkspaceDir = await tempDirs.make("openclaw-experience-worktree-");
+    const skillDir = path.join(canonicalWorkspaceDir, "skills", "deployment-preflight");
+    await writeSkill({
+      dir: skillDir,
       name: "deployment-preflight",
       description: "Check deployment prerequisites before retrying.",
-      proposal_content: "# Deployment Preflight\n\nOperator-authored preflight steps.\n",
+      body: "# Deployment Preflight\n\nOperator-authored preflight steps.\n",
     });
-    await seedTool.execute("seed-apply", {
-      action: "apply",
-      proposal_id: (seeded.details as { id: string }).id,
-      reason: "seed live skill",
+    await fs.cp(skillDir, path.join(worktreeWorkspaceDir, "skills", "deployment-preflight"), {
+      recursive: true,
     });
 
     runEmbeddedAgent.mockImplementation(async (params) => {
@@ -137,28 +136,47 @@ describe("experience review auto apply", () => {
         agentId: "main",
         runId: "foreground-run",
         sessionKey: "agent:main:main",
-        workspaceDir,
+        workspaceDir: worktreeWorkspaceDir,
         modelProviderId: "openai",
         modelId: "gpt-test",
       },
-      config: { skills: { workshop: { autonomous: { mode: "auto" } } } },
+      config: {
+        agents: { list: [{ id: "main", default: true, workspace: canonicalWorkspaceDir }] },
+        skills: { workshop: { autonomous: { mode: "auto" as const } } },
+      },
       transcript: "[user]\nRefine the deployment workflow.",
       modelIterations: 10,
     };
 
-    await runSkillExperienceReview(candidate, {
-      getCurrentConfig: () => candidate.config ?? {},
-    });
+    await runWithCanonicalSkillWorkspace(canonicalWorkspaceDir, () =>
+      runSkillExperienceReview(candidate, {
+        getCurrentConfig: () => candidate.config ?? {},
+      }),
+    );
 
-    const manifest = await listSkillProposals({ workspaceDir });
+    const manifest = await listSkillProposals({
+      agentId: "main",
+      workspaceDir: canonicalWorkspaceDir,
+    });
     const updateEntry = manifest.proposals.find((entry) => entry.kind === "update");
     expect(updateEntry).toMatchObject({
       skillKey: "deployment-preflight",
       status: "applied",
     });
+    const inspected = await inspectSkillProposal(updateEntry?.id ?? "", {
+      agentId: "main",
+      workspaceDir: canonicalWorkspaceDir,
+    });
+    expect(inspected?.record.target.skillFile).toBe(path.join(skillDir, "SKILL.md"));
+    await expect(fs.readFile(path.join(skillDir, "SKILL.md"), "utf8")).resolves.toContain(
+      "Reviewer-rewritten steps.",
+    );
     await expect(
-      fs.readFile(`${workspaceDir}/skills/deployment-preflight/SKILL.md`, "utf8"),
-    ).resolves.toContain("Reviewer-rewritten steps.");
+      fs.readFile(
+        path.join(worktreeWorkspaceDir, "skills", "deployment-preflight", "SKILL.md"),
+        "utf8",
+      ),
+    ).resolves.toContain("Operator-authored preflight steps.");
   });
 
   it("auto-applies reviewer patch proposals composed from the live body", async () => {
